Start shells in dedicated scopes
This puts every single shell into a dedicated scope which isolates
shells from the parent WezTerm process and puts every shell under
separate CPU and memory accounting.

Practically this means that if a program in a shell consumes an
excessive amount of memory systemd-oomd takes down just that shell
instead of the entire wezterm process.