Get rid of kmail
KMail is a terrible mail client from the last century.  It still doesn't
support HTML mails properly (e.g. replying with quoted HTML content
totally destroys the HTML layout of the quoted content), and makes way
too big a fuzz about HTML in mails in general.  This is not 2000 anymore.