Add custom time formats to date menu button
Do not add an additional indicator; instead fiddle with the date menu
display to get our custom formats into the button.

This makes the position of the clock more predictable.

As a side-benefit we create a lot less types; in particular we also
reuse the wall clock of the date menu button.