Add protected 1password repo
Move 1password repo to a separate tree to quickly enable/disable it,
and encrypt the repository file to hide the password.

In install.bash manually decrypt the file after applying all trees.

We encrypt with gpg because by means of a pacman dependency gpg is
always present on any pacstrapped Arch system.